// motorbike_be/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleName = {
  ADMIN: 'ADMIN',
  RECEPTIONIST: 'RECEPTIONIST',
  TECHNICIAN: 'TECHNICIAN'
} as const

export type RoleName = (typeof RoleName)[keyof typeof RoleName]


export const VehicleType = {
  MANUAL: 'MANUAL',
  SCOOTER: 'SCOOTER',
  BIG: 'BIG'
} as const

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType]


export const AppointmentStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  COMPLETED: 'COMPLETED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]


export const RepairOrderStatus = {
  RECEIVED: 'RECEIVED',
  IN_PROGRESS: 'IN_PROGRESS',
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED'
} as const

export type RepairOrderStatus = (typeof RepairOrderStatus)[keyof typeof RepairOrderStatus]


export const PaymentMethod = {
  CASH: 'CASH',
  BANK_TRANSFER: 'BANK_TRANSFER',
  QR_CODE: 'QR_CODE'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const VoucherStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  REVOKED: 'REVOKED'
} as const

export type VoucherStatus = (typeof VoucherStatus)[keyof typeof VoucherStatus]


export const BlogStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus]


export const ReviewStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  HIDDEN: 'HIDDEN'
} as const

export type ReviewStatus = (typeof ReviewStatus)[keyof typeof ReviewStatus]


export const NotificationType = {
  APPOINTMENT_CONFIRM: 'APPOINTMENT_CONFIRM',
  APPOINTMENT_REMINDER: 'APPOINTMENT_REMINDER',
  REPAIR_COMPLETED: 'REPAIR_COMPLETED',
  MAINTENANCE_DUE: 'MAINTENANCE_DUE',
  OIL_CHANGE_DUE: 'OIL_CHANGE_DUE',
  STOCK_ALERT: 'STOCK_ALERT',
  PRICE_APPROVAL: 'PRICE_APPROVAL'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]